+++ src/Cooked/MockChain.hs
@@ -4,10 +4,14 @@ module Cooked.MockChain (module X) where
 
 import Cooked.MockChain.AutoFilling as X
 import Cooked.MockChain.Balancing as X
-import Cooked.MockChain.BlockChain as X
+import Cooked.MockChain.Common as X
 import Cooked.MockChain.Direct as X
+import Cooked.MockChain.Error as X
+import Cooked.MockChain.Instances as X
+import Cooked.MockChain.Misc as X
 import Cooked.MockChain.MockChainState as X
-import Cooked.MockChain.Staged as X
+import Cooked.MockChain.Read as X
 import Cooked.MockChain.Testing as X
 import Cooked.MockChain.UtxoSearch as X
 import Cooked.MockChain.UtxoState as X
+import Cooked.MockChain.Write as X
